bot v.2/requirements.txt
aiogram==2.25.2
nest_asyncio
numpy
matplotlib
yfinance
pandas
scikit-learn
xgboost
psycopg2-binary